fix some moderation command can be use in user with higher role

+++ slashcommands/moderation/ban.js
@@ -1,10 +1,15 @@
 const run = async (client, interaction) => {
     const member = interaction.options.getMember('user');
-    const reason = interaction.options.getString('reason'); 
-    member.ban(reason);
-    interaction.deferReply();
-    interaction.deleteReply();
-	interaction.channel.send(`Tài khoản ${member} đã bị cấm với lý do ${reason}`).catch((err) => {console.log(err)})
+    const reason = interaction.options.getString('reason');
+    if(interaction.user.roles.highest.position > member.user.roles.highest.position) {
+        member.ban(reason);
+        interaction.deferReply();
+        interaction.deleteReply();
+        interaction.channel.send(`Tài khoản ${member} đã bị cấm với lý do ${reason}`).catch((err) => {console.log(err)})
+    }
+    else {
+        interaction.reply({ content: `Bạn không thể thực hiện hàng động này với ${member}`, ephemeral: true }).catch((err) => {console.log(err)})
+    }
 }
 module.exports = {
 	name: "ban",
